requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0.0
httpx>=0.24.0
watchdog>=3.0.0
